@@ -1,62 +1,38 @@
 Rails.application.routes.draw do
-  get 'landing/index'
-
+  root to: 'visitors#index'
   devise_for :users
-  # The priority is based upon order of creation: first created -> highest priority.
-  # See how all your routes lay out with "rake routes".
+  # resources :users
 
-  # You can have the root of your site routed with "root"
-  # root 'welcome#index'
+  # Admin namespace
+  namespace :admin, path: 'admin' do
+    resources :users
+    # Admin root path
+    match '/', to: 'dashboard#index', via: [:get], as: :root
+  end
 
-  # Example of regular route:
-  #   get 'products/:id' => 'catalog#view'
+  # SubAdmin namespace
+  namespace :subadmin, path: 'sub' do
+    resources :users
+    # SubAdmin root path
+    match '/', to: 'dashboard#index', via: [:get], as: :root
+  end
 
-  # Example of named route that can be invoked with purchase_url(id: product.id)
-  #   get 'products/:id/purchase' => 'catalog#purchase', as: :purchase
+  # Manager namespace
+  namespace :manager, path: 'm' do
+    # Manager root path
+    match '/', to: 'dashboard#index', via: [:get], as: :root
+  end
 
-  # Example resource route (maps HTTP verbs to controller actions automatically):
-  #   resources :products
+  # User namespace
+  namespace :user, path: 'u' do
+    # User root path
+    match '/', to: 'dashboard#index', via: [:get], as: :root
+  end
 
-  # Example resource route with options:
-  #   resources :products do
-  #     member do
-  #       get 'short'
-  #       post 'toggle'
-  #     end
-  #
-  #     collection do
-  #       get 'sold'
-  #     end
-  #   end
+  # Guest namespace
+  namespace :guest, path: 'g' do
+    # Guest root path
+    match '/', to: 'dashboard#index', via: [:get], as: :root
+  end
 
-  # Example resource route with sub-resources:
-  #   resources :products do
-  #     resources :comments, :sales
-  #     resource :seller
-  #   end
-
-  # Example resource route with more complex sub-resources:
-  #   resources :products do
-  #     resources :comments
-  #     resources :sales do
-  #       get 'recent', on: :collection
-  #     end
-  #   end
-
-  # Example resource route with concerns:
-  #   concern :toggleable do
-  #     post 'toggle'
-  #   end
-  #   resources :posts, concerns: :toggleable
-  #   resources :photos, concerns: :toggleable
-
-  # Example resource route within a namespace:
-  #   namespace :admin do
-  #     # Directs /admin/products/* to Admin::ProductsController
-  #     # (app/controllers/admin/products_controller.rb)
-  #     resources :products
-  #   end
-
-  # This is where I started... Added the following line as per devise install instructions
-  root to: "landing#index"
 end
